\name{wavFeatExt-package}
\alias{wavFeatExt-package}
\alias{wavFeatExt}
\docType{package}
\title{
\packageTitle{wavFeatExt}
}
\description{This package contains utility functions to perform non-decimated Haar wavelet transformation and managing the extracted features for prediction.
\packageDescription{wavFeatExt}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{wavFeatExt}
\packageIndices{wavFeatExt}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{wavFeatExt}

Maintainer: \packageMaintainer{wavFeatExt}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
